enhancement : Post 좋아요 및 댓글 개수 구하는 로직 개선 , Dto 유효성 검증

+++ dplanner/src/test/java/com/dp/dplanner/repository/PostMemberLikeRepositoryTest.java
@@ -53,7 +53,7 @@ public class PostMemberLikeRepositoryTest {
 
         postMemberLikeRepository.saveAll(Arrays.asList(like1, like2));
 
-        int likeCount = postMemberLikeRepository.countDistinctByPostId(post.getId());
+        Long likeCount = postMemberLikeRepository.countDistinctByPostId(post.getId());
 
         assertThat(likeCount).isEqualTo(2);
     }
